Extracted superclass RandomizedShapeTest from AbstractTestShapes.
